introduce pluginTester function for shared code
