+++ storybook-addon-export-to-codesandbox/src/plugins/fullsource.test.ts
@@ -1,18 +1,4 @@
-import pluginTester from 'babel-plugin-tester';
-import * as path from 'path';
 import plugin, { PLUGIN_NAME } from './fullsource';
+import pluginTester from './pluginTester';
 
-const fixturesDir = path.join(__dirname, `__fixtures__/${PLUGIN_NAME}`);
-
-const defaultDependencyReplace = { replace: '@fluentui/react-components' };
-
-pluginTester({
-  pluginOptions: {
-    '@fluentui/react-button': defaultDependencyReplace,
-    '@fluentui/react-menu': defaultDependencyReplace,
-    '@fluentui/react-link': defaultDependencyReplace,
-  },
-  pluginName: PLUGIN_NAME,
-  plugin,
-  fixtures: fixturesDir,
-});
+pluginTester(plugin, PLUGIN_NAME);
