+++ storybook-addon-export-to-codesandbox/src/plugins/modifyImports.test.ts
@@ -1,17 +1,6 @@
-import pluginTester from 'babel-plugin-tester';
-import * as path from 'path';
 import plugin, { PLUGIN_NAME } from './modifyImports';
+import pluginTester from './pluginTester';
 
-const defaultDependencyReplace = { replace: '@fluentui/react-components' };
-const fixturesDir = path.join(__dirname, `__fixtures__/${PLUGIN_NAME}`);
-pluginTester({
-  pluginOptions: {
-    '@fluentui/react-button': defaultDependencyReplace,
-    '@fluentui/react-menu': defaultDependencyReplace,
-    '@fluentui/react-link': defaultDependencyReplace,
-    '@fluentui/react-unstable-component': { replace: '@fluentui/react-components/unstable' },
-  },
-  pluginName: PLUGIN_NAME,
-  plugin,
-  fixtures: fixturesDir,
+pluginTester(plugin, PLUGIN_NAME, {
+  '@fluentui/react-unstable-component': { replace: '@fluentui/react-components/unstable' },
 });
